Updated getturbotx
Updated getturbotx. Fixed all errors, and added a block status (if its in block or not) and how many delegates validated it. Also added a backup check because it might go into a block before the person checks it. Also added verification to all network data nodes after DBFT check as this will be faster and work request.

leaving in the debug messages, and will remove in another commit so I can readd them if needed