change id and verify sender
Change the id to first 5 of tx hash. Verify that sender is someone in the tx unspents. Added better error checking

These changes are due to since were not using DPOPS delegates to store the data (using a centralized db due to speed of commands for this type of protocol) we need to setup the data as all L1 blockchain and verifiable data, this way one knows the db cant be edited as either the data could not work if faked or would be invalid as its all L1 blockchain data

the timestamp is the only thing to maybe change at this point