Change timestamp
Change timestamp, so it uses calculated one when less than a block time, but the blockchain L1 data after added to a block. This will prevent people from tying to modify the timestamp as a wait in line after a purchase would usually take more than 5 minutes, and they would probably be remembered. This was mostly to prevent people from coming back hours later.